admin_ui/requirements.txt
fastapi
uvicorn
mysql-connector-python
